requirements.txt
appnope==0.1.2
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
backports.shutil-get-terminal-size==1.0.0
batchgenerators==0.21
bcrypt==3.2.0
bleach==3.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
cryptography==3.4.7
cycler==0.10.0
decorator==4.4.2
defusedxml==0.7.1
dicom2nifti==2.2.12
dill==0.3.4
entrypoints==0.3
future==0.18.2
idna==2.10
imageio==2.9.0
imagezmq==1.1.1
imutils==0.5.4
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
linecache2==1.0.0
MarkupSafe==1.1.1
matplotlib==3.4.1
MedPy==0.4.0
mistune==0.8.4
multiprocess==0.70.12.2
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
nibabel==3.2.1
-e git+https://github.com/MIC-DKFZ/nnUNet.git@627c2473a0c3372e39c8eb27f28fdd82d46831df#egg=nnunet
notebook==6.3.0
numpy==1.20.2
opencv-python==4.5.3.56
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
paramiko==2.7.2
parso==0.8.2
pathos==0.2.8
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pox==0.3.0
ppft==1.6.6.4
prometheus-client==0.10.1
prompt-toolkit==3.0.18
ptyprocess==0.7.0
pycparser==2.20
pydicom==2.1.2
Pygments==2.8.1
PyNaCl==1.4.0
pynrrd==0.4.2
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
pyzmq==22.0.3
reprint==0.6.0
requests==2.25.1
scikit-image==0.18.1
scikit-learn==0.24.2
scipy==1.6.3
scp==0.13.6
Send2Trash==1.5.0
SimpleITK==1.2.4
six==1.15.0
sklearn==0.0
terminado==0.9.4
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2021.4.8
torch==1.8.1
torchvision==0.9.1
tornado==6.1
tqdm==4.60.0
traceback2==1.4.0
traitlets==5.0.5
typing-extensions==3.7.4.3
unittest2==1.1.0
urllib3==1.26.4
vtk==9.0.1
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
